Set last-modified header on messages
This should at least hint Google and others about how often our pages
really do change. Only set it for messages for now, since other pages
are excluded from cache anyway.

Last modified is set to the lasat modified entr in the database, we don't
make any provisions for changes in the templates or the rendering itself.
We might want to consider this in the future, maybe looking at the datestamp
of the template file and the view.py, or something along that line.

Implement simple If-Modified-Since checks, even though that's not going
to work anyway with the current version of Varnish - but hopefully it
will with the next one.