Implement check functionality for reparsing messages
By default, reparse_message.py will now reparse the message and print the
diff, instead of actually updating it. Give it a --update to make it
actually do the update, like before. This makes it a good tool for testing
updates to the message parsing code.

Also implement --all and --sample <n>, to have it run the diff across
all or a certain number of messages in the archives. (Be *very* careful
about running it with both --all and --update..)