Add script to purge emails from varnish
This is typically needed when manually editing a message such as setting
the hidden reason.

We should perhaps extend this at some point to support doing the hiding
itself, but for now that is still done manually and you just pass the
messageid to this script to clear it from the varnish caches.