Filter out hidden messages in list of messages as well
There is no reason to include a link to a page that just says "this
message isn't there". And sometimes the spam is found in the subject as
well.

This should probably have been done as part of
d5df0573710c5e5bf562da193c43e650cb070998, but better late than never.