Generate mbox data from database instead of cached files
Previously we only did this for thread-mbox downloads, but since the
move to pglister we need to do it for the main mbox downloads as well.

In order to make this more efficient, ensure that we use a server-side
cursor when getting the messages. We already sent them back to the
webserver in streaming mode, but not between the database and django.