Implement per-list permissions
This assumes we sync subscribers over from the list server (using
pglister), getting their community authentication usernames. Then, by
requesting a community auth login, it's possible to restrict the session
to view only those lists the user is subscribed to.

To view emails, the user must be subscribed to *all* the lists that the
thread the message belongs to has shown up. This means that messages can
dissappear from a listing if somebody CCs in a higher security level
list.

NOTE! After installing this code, the PUBLIC_ARCHIVES setting must be
set to True to retain previous behaviour!

Reviewed by Stephen Frost