Use list name, not list id, in generated search form
The coupling between main site and archives was supposed to be loose,
which means that id's can be different. Meaning a listid sent from the
archives server might mean a different list on the website. So make sure
we pass out the list *name*, which the website knows what to do with.