Don't do a #-redirect for first message in thread
On the view-whole-thread page we previously did a redirect back to the
anchor of the very first mail of the thread when loading the first mail
of the thread. This pollutes the URL more than necessary.

We still do the redirect for messages that are not the first, to make it
possible to link to an individual message.