Make flat messageview redirect to md5 of messageid
Redirecting to the full messageid pollutes the URL pretty badly. md5 is
not the best of hashes, but the likelihood of getting an md5 collission
on the messageid *within a single thread* is extremely low -- and of
course if it happens the only effect is on the initial scroll of the
window.