Switch to using tidylib rather than tidy
tidylib (http://countergram.github.io/pytidylib/) is maintained,
the old tidy one (https://cihar.com/software/utidylib/) is not. And in
particular, python3 support is in the new one.

Generates some minor changs in the existing archives, but it seems to be
just whitespace and some actual incorrectness in the old output.