Pin the navigation bar to the top of the page.
From user feedback, the scrolling navigation bar would cause problems with
readability, especially in larger font views. This keeps the navigation bar
pinned to the top, until we can find a way to more easily handle this.

In the context of pgarchives, this is one bit of "future-proofing" as we do not
currently show the shout box in pgarchives, but may do so in the near future.