Extra quote the url from 'next' in authentication
If the next URL contained a + django would helpfully change that into a
space, which broke things further down the chain. So put back this
escaping, in the hope that the redirect will be correct down the road.

This fixes is for spaces, let's hope it doesn't instead break it for
something else.