Use article:author instead of og:article:author
The specification is a bit unclear on which it should be and completely
unclear about what should go in the field, but at least it seems most
implementations use article:author rather than og:article:author, so try
changing it.

Also change og:published_time to be article:published_time.