Properly quote URL in link to resend an email
The lack of this would result in a 404 for any users who clicked a
resend link when (1) they were not logged in, and (2) the messageid
cotnained a plus character. This would then end up getting unescaped
one too many times in the authentication flow and came out as a space on
the other end instead of a plus.

Reported by Justin Pryzby (and several others, but Justin was persistant
in tracking down good examples)