Clean up accidentally commited files
Back in 2018, commit 4d159ca accidentally included unrelated
functionality which was not completed. This causes the database to be
out of sync with the models defined in the code.

This commit reverts those parts that were not supposed to be included
and leaves the changes that actually were. The code can be re-added once
completed...

Spotted by Célestin Matte