Round of since/ and before/ URLs to whole days
Previously we'd generate links for each individual email, when clicking
the link on the emails. This generates a huge number of URLs that
contain basically the same thing, which has an adverse effect on both
caching and (stupid) bots. So round it off to even days which puts at
least some level of a cap on it.

Also, when a hit comes in that specifies the full hour, redirect it back
to the rounded-off value.